Refactor of ChartUtil.
