+++ src/components/Utilites/ChartUtil.js
@@ -76,4 +76,11 @@ function setOddPercentsToDisplayNone(ratingPercent) {
   }
 }
 
-module.exports = { TickChange };
+const TimelineOptions = [
+  { value: 'all', label: 'All Available' },
+  { value: '30', label: 'Last 30 Days' },
+  { value: '60', label: 'Last 60 Days' },
+  { value: 'YTD', label: 'Year to Date' },
+];
+
+module.exports = { TickChange, TimelineOptions };
